添加md5检查，不存在才上传 多文件上传机制
添加空格

还有多个bug，重复上传文件，但是hash是判断正确的，删除文件时把真实的文件给删除了，应该只删除对应的记录

简单的手动实现了将session存储在redis中
下一步将使用尝试spring session

实现了策略模式demo

v1.1.2
采用文件引用，实现了虚拟删除文件记录，修改部分数据表的设计

+++ src/main/java/com/example/controller/FileInfoController.java
@@ -30,16 +30,16 @@ public class FileInfoController {
         return Result.success(info);
     }
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/{id}")  //file_info 的id
     public Result delete(@PathVariable Long id, HttpServletRequest request) {
         Account account = (Account) request.getSession().getAttribute("user");
         FileInfo info = fileInfoService.findById(id);
         if (!account.getLevel().equals(info.getLevel()) || !account.getId().equals(info.getUploadUserId())) {
             return Result.error("1001", "不能删除他人的记录");
         }
-        fileInfoService.delete(id);
-        // 删除对应文件记录
-        if (info.getFileId() != null) {
+        int value = fileInfoService.delete(id);
+        // 删除对应真实文件
+        if (info.getFileId() != null && value == 0) {
             nxSystemFileController.deleteFile(info.getFileId().toString());
         }
         return Result.success();
